feat(skills): add Algolia skill and update skill data structure; include duration notes in translations
