+++ apps/front/src/components/skills/skill-card.tsx
@@ -3,7 +3,8 @@
 import { useTranslations } from 'next-intl'
 import { useEffect, useRef, useState } from 'react'
 
-import { Skill } from './skills-data'
+import { Skill } from '@/types/skill'
+
 import { ButtonShimmer } from '../shared/button-shimmer'
 
 interface Props {
@@ -71,8 +72,12 @@ export function SkillCard({ skill }: Props) {
             {t(`${skill.name}.description`)}
           </p>
         </div>
-        <span className="bg-muted-foreground/70 text-secondary rounded-full px-3 py-1 text-base font-medium backdrop-blur-md">
-          {skill.proficiency}%
+        <span className="bg-muted-foreground/70 text-secondary rounded-full px-3 py-1 text-base font-medium text-nowrap backdrop-blur-md">
+          {skill.months >= 12
+            ? skill.months == 18
+              ? t('moreThanOneYear')
+              : t('years', { count: Math.floor(skill.months / 12) })
+            : t('months', { count: skill.months })}
         </span>
       </div>
       <div className="bg-muted absolute top-0 left-0 z-[1] h-full w-full rounded-2xl" />
