+++ apps/front/src/components/skills/skills-data.tsx
@@ -1,20 +1,14 @@
 import { SquareTerminal } from 'lucide-react'
 import Image from 'next/image'
 
-export interface Skill {
-  name: string
-  description: string
-  proficiency: number
-  color: string
-  icon: React.ReactNode
-}
+import { Skill } from '@/types/skill'
 
 export const skillsData: Skill[] = [
   {
     name: 'python',
     description: 'Building robust backend systems and data pipelines',
     proficiency: 90,
-    color: 'rgba(218, 165, 32, 0.6)',
+    months: 24,
     icon: (
       <Image
         src="/icons/skills/Python.svg"
@@ -29,7 +23,7 @@ export const skillsData: Skill[] = [
     name: 'typeScript',
     description: 'Building high-performance full-stack web applications',
     proficiency: 85,
-    color: 'rgba(0, 0, 139, 0.6)',
+    months: 18,
     icon: (
       <Image
         src="/icons/skills/TypeScript.svg"
@@ -44,7 +38,7 @@ export const skillsData: Skill[] = [
     name: 'postgreSQL',
     description: 'Building complex, scalable and efficient databases',
     proficiency: 75,
-    color: 'rgba(30, 144, 255, 0.6)',
+    months: 12,
     icon: (
       <Image
         src="/icons/skills/PostgreSQL.svg"
@@ -58,8 +52,8 @@ export const skillsData: Skill[] = [
   {
     name: 'nestJs',
     description: 'Building high-performance backend applications',
-    proficiency: 70,
-    color: 'rgba(220, 38, 57, 0.6)',
+    proficiency: 75,
+    months: 12,
     icon: (
       <Image
         src="/icons/skills/NestJS.svg"
@@ -74,7 +68,7 @@ export const skillsData: Skill[] = [
     name: 'nextJs',
     description: 'Building high-performance frontend applications',
     proficiency: 70,
-    color: 'rgba(255, 255, 255, 0.1)',
+    months: 8,
     icon: (
       <Image
         src="/icons/skills/NextJs.svg"
@@ -88,8 +82,8 @@ export const skillsData: Skill[] = [
   {
     name: 'ciCd',
     description: 'Automating workflows and tasks',
-    proficiency: 60,
-    color: 'rgba(122, 34, 255, 0.6)',
+    proficiency: 65,
+    months: 6,
     icon: (
       <Image
         src="/icons/GitHub.svg"
@@ -103,8 +97,8 @@ export const skillsData: Skill[] = [
   {
     name: 'lowLevel',
     description: 'Scripting and low-level programming',
-    proficiency: 55,
-    color: 'rgba(122, 34, 255, 0.6)',
+    proficiency: 60,
+    months: 4,
     icon: <SquareTerminal width={70} height={70} className="text-secondary" />
   }
 ]
