+++ apps/front/src/types/skill.ts
@@ -1,0 +1,7 @@
+export interface Skill {
+  name: string
+  description: string
+  proficiency: number
+  months: number
+  icon: React.ReactNode
+}
